Local Route Set: add refeence to LocalRoute State Changes
resolves
(There wasn’t any text above and it’s not clear how this is enforced.  I have an idle route and an unconfirmed route, the unconfirmed route is updated to confirmed, I now have two valid route entries and no way to demote/decide or otherwise purge one of them but I’m not allowed to do that.  Setting the worse and previously idle route to unconfirmed seems wrong and broken. JWD!)