Regeneration-> Forwarding (except with RERR)
Change the language from regeneration to forwarding. This does *not* include an update for the sec cons (that's for another commit).
Note that I didn't change the wording when it comes to RERRs (yet) as they only travel one hop. I think I recall a conversation on that on [manet], but I've been unable to find it... :(